changes for code climate
