+++ app/src/main/java/org/ole/planet/takeout/Dashboard.java
@@ -42,15 +42,12 @@ import org.ole.planet.takeout.utilities.Utilities;
 import java.util.ArrayList;
 
 
-public class Dashboard extends AppCompatActivity {
+public class Dashboard extends DashboardElements {
     private Drawer result = null;
     private Toolbar mTopToolbar;
     AccountHeader headerResult;
     public static final String MESSAGE_PROGRESS = "message_progress";
     private static final int PERMISSION_REQUEST_CODE = 111;
-
-    private EditText feedbackText;
-    private RadioGroup choice1, choice2;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -212,31 +209,6 @@ public class Dashboard extends AppCompatActivity {
         dialog.show();
     }
 
-    public void disableSubmit(MaterialDialog dialog) {
-        final View submitButton = dialog.getActionButton(DialogAction.POSITIVE);
-        submitButton.setEnabled(false);
-        feedbackText = dialog.getCustomView().findViewById(R.id.user_feedback);
-        choice1 = dialog.getCustomView().findViewById(R.id.choice1);
-        choice2 = dialog.getCustomView().findViewById(R.id.choice2);
-        feedbackText.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-                if (choice1.getCheckedRadioButtonId() == -1 || choice2.getCheckedRadioButtonId() == -1 || s.toString().trim().length() == 0) {
-                    submitButton.setEnabled(false);
-                } else {
-                    submitButton.setEnabled(true);
-                }
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-            }
-        });
-    }
 
 
     public void openCallFragment(Fragment newfragment) {
